# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bayyina CRM API
  description: |
    REST API for Bayyina CRM — an education center management system.

    ## Authentication
    All protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are obtained via `POST /api/auth/login` or `POST /api/auth/verify-otp`.
    Tokens expire after **30 days**.

    ## Roles
    | Role      | Description                                      |
    |-----------|--------------------------------------------------|
    | `admin`   | Full access to all resources                     |
    | `teacher` | Read access to students; manage groups/payments  |
    | `student` | Read-only access to own data                     |

    ## Error Format
    All errors follow this structure:
    ```json
    { "code": "errorCode", "message": "Human readable message" }
    ```

  version: 1.0.0
  contact:
    name: Mardonbek Khamidov
    email: mardonbek@bayyina.uz
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.bayyina.uz/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued on login or OTP verification. Valid for 30 days.

  # ─────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────
  schemas:
...
    LeadSource:
      type: object
      description: A configurable lead acquisition channel (e.g. Telegram, Instagram).
      properties:
        _id:
          type: string
          example: "664f1a2b3c4d5e6f7a8b9c11"
        name:
          type: string
          description: Display name of the source
          example: "Telegram"
        slug:
          type: string
          description: Lowercase unique identifier
          example: "telegram"
        createdBy:
          oneOf:
            - type: string
              nullable: true
            - type: object
              properties:
                _id: { type: string, example: "664f1a2b3c4d5e6f7a8b9c0f" }
                firstName: { type: string, example: "Sarvinoz" }
                lastName: { type: string, example: "Yusupova" }
        createdAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"

    # ── CourseType ───────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "invalidCredentials"
        message:
          type: string
          example: "Telefon raqam yoki parol noto'g'ri"

    # ── Record ──────────────────────────────────
...
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Number of results per page (default 20, max 100)
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            tokenMissing:
              summary: No token provided
              value: { code: "tokenMissing", message: "Token mavjud emas" }
            tokenExpired:
              summary: Token has expired
              value: { code: "tokenExpired", message: "Token muddati tugagan" }
            invalidToken:
              summary: Token is malformed
              value: { code: "invalidToken", message: "Yaroqsiz token kiritildi" }
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "forbidden"
            message: "Ruxsat berilmadi"
...
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded (120 req / 3 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "tooManyRequests"
            message: "Juda ko'p so'rovlar. Iltimos, 3 daqiqadan keyin urinib ko'ring"
...
paths:
...
  /lead-sources:
    get:
      tags: [Lead Sources]
      summary: List lead sources
      description: Returns a paginated list of lead sources. Supports free-text search (`q`). **Admin only.**
      operationId: getLeadSources
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - name: q
          in: query
          description: Search by name or slug
          schema:
            type: string
            example: "telegram"
      responses:
        "200":
          description: Lead sources returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadSources:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeadSource"
                  total: { type: integer, example: 5 }
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  totalPages: { type: integer, example: 1 }
                  hasNextPage: { type: boolean, example: false }
                  hasPrevPage: { type: boolean, example: false }
                  code: { type: string, example: leadSourcesFound }
                  message: { type: string, example: "Manbalar ro'yxati olindi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      tags: [Lead Sources]
      summary: Create a lead source
      description: Creates a new lead source. **Admin only.**
      operationId: createLeadSource
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Telegram"
                slug:
                  type: string
                  example: "telegram"
      responses:
        "201":
          description: Lead source created
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadSource:
                    $ref: "#/components/schemas/LeadSource"
                  code: { type: string, example: leadSourceCreated }
                  message: { type: string, example: "Manba muvaffaqiyatli yaratildi" }
        "400":
          description: Missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "missingField", message: "Manba nomi kiritilishi shart" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /lead-sources/{id}:
    get:
      tags: [Lead Sources]
      summary: Get a lead source by ID
      operationId: getLeadSource
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c11"
      responses:
        "200":
          description: Lead source found
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadSource:
                    $ref: "#/components/schemas/LeadSource"
                  code: { type: string, example: leadSourcesFound }
                  message: { type: string, example: "Manbalar ro'yxati olindi" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "leadSourceNotFound", message: "Manba topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      tags: [Lead Sources]
      summary: Update a lead source
      operationId: updateLeadSource
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c11"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Instagram"
                slug:
                  type: string
                  example: "instagram"
      responses:
        "200":
          description: Lead source updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadSource:
                    $ref: "#/components/schemas/LeadSource"
                  code: { type: string, example: leadSourceUpdated }
                  message: { type: string, example: "Manba muvaffaqiyatli yangilandi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "leadSourceNotFound", message: "Manba topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags: [Lead Sources]
      summary: Delete a lead source
      operationId: deleteLeadSource
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c11"
      responses:
        "200":
          description: Lead source deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string, example: leadSourceDeleted }
                  message: { type: string, example: "Manba muvaffaqiyatli o'chirildi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "leadSourceNotFound", message: "Manba topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ══════════════════════════════════════════
  # COURSE TYPES
  # ══════════════════════════════════════════
...